# 3° Anno/Programmazione per il Web/Progetto/WASAText/doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000"
info:
  title: WASAText
  description: una piattaforma di messaggistica che supporta chat individuali e di gruppo, accessibile tramite PC
  version: 1.0.0
...
paths:
...
  /chats:
  # Ottengo tutte le chat dell'utente autenticato (in ordine cronologico inverso)
    get:
      parameters:
        - $ref: '#/components/parameters/Ordinamento'
      tags:
        - chat
      summary: Ottieni tutte le chat dell'utente autenticato
      description: Restituisce la lista di tutte le chat (individuali e di gruppo) dell'utente autenticato.
      operationId: getMyConversations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Elenco chat recuperato con successo
          content:
            application/json:
              schema:
                type: array
                description: Lista di chat dell'utente
                items:
                  $ref: '#/components/schemas/Chat'
                minItems: 0
                maxItems: 1000
        '404':
          $ref: '#/components/responses/NotFound'
        '500': 
          $ref: '#/components/responses/InternalServerError'

    # Creazione di una nuova chat individuale o di gruppo
    post:
      tags:
        - chat
      summary: Crea una nuova chat
      description: |-
        Permette di creare una nuova chat individuale o di gruppo.
        Restituisce l'oggetto chat creato.
      operationId: createConversation
      security:
        - bearerAuth: []
      requestBody:
        description: Dettagli della nuova chat
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatRequest'
      responses:
        '201':
          description: Chat creata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # SEZIONE 2.1 - Chat specifica
  /chats/{chat_id}:
    parameters:
      - name: chat_id
        in: path
        description: ID univoco della chat.
        required: true
        schema:
            type: string
            minLength: 1
            maxLength: 64
            pattern: '^[a-zA-Z0-9_-]{1,64}$'
    get:
      tags:
        - chat
      summary: Ottieni i dettagli di una chat specifica
      description: Restituisce i dettagli della chat identificata da chat_id.
      operationId: getConversation
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dettagli chat recuperati con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - gruppo
      summary: Imposta o aggiorna il nome del gruppo  
      description: |-
        Permette di impostare o aggiornare il nome di una chat di gruppo.
        Restituisce l'oggetto chat di gruppo aggiornato.
      operationId: setGroupName
      security:
        - bearerAuth: []
      requestBody:
        description: Nuovo nome del gruppo
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetGroupNameRequest'
      responses:
        '200':
          description: Nome del gruppo aggiornato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - chat
      summary: Elimina una chat specifica
      description: |-
        Permette di eliminare una chat specifica (individuale o di gruppo).
      operationId: deleteConversation
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Chat eliminata con successo
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  
  # SEZIONE 2.2 - Impostare o aggiornare la foto del gruppo
  /chats/{chat_id}/photo:
    parameters:
        - name: chat_id
          in: path
          description: ID univoco della chat.
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
            pattern: '^[a-zA-Z0-9_-]{1,64}$'
    put:
      tags:
        - gruppo
      description: |-
        Permette di impostare o aggiornare la foto di una chat di gruppo.
        Restituisce l'oggetto chat di gruppo aggiornato.
      summary: Imposta o aggiorna la foto del gruppo
      operationId: setGroupPhoto
      security:
        - bearerAuth: []
      requestBody:
        description: Nuova foto del gruppo
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetGroupPhotoRequest'
      responses:
        '200':
          description: Foto aggiornata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # SEZIONE 2.3 - Impostare o aggiornare la descrizione del gruppo
  /chats/{chat_id}/description:
    parameters:
        - name: chat_id
          in: path
          description: ID univoco della chat.
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
            pattern: '^[a-zA-Z0-9_-]{1,64}$'
    put:
      tags:
        - gruppo
      description: |-
        Permette di impostare o aggiornare la descrizione di una chat di gruppo.
        Restituisce l'oggetto chat di gruppo aggiornato.
      summary: Imposta o aggiorna la descrizione del gruppo
      operationId: setGroupDescription
      security:
        - bearerAuth: []
      requestBody:
        description: Nuova descrizione del gruppo
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetGroupDescriptionRequest'
      responses:
        '200':
          description: Descrizione aggiornata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # SEZIONE 2.2 - Aggiunta partecipanti ad una chat di gruppo
  /chats/{chat_id}/participants:
    parameters:
        - name: chat_id 
          in: path
          description: ID univoco della chat.
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
            pattern: '^[a-zA-Z0-9_-]{1,64}$'
    post: 
      tags:
        - gruppo
      summary: Aggiungi uno o più membri al gruppo
      description: |- 
        Permette all'utente (solo se è definito admin) di aggiungere uno
        o più utenti al gruppo 
      operationId: addToGroup 
      security: 
        - bearerAuth: []
      requestBody:
        description: Elenco degli ID degli utenti da aggiungere 
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddParticipantsRequest'
      responses:
        '200':
          description: Membri aggiunti con successo
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/Chat'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # SEZIONE 2.3 - Utente lascia il gruppo di appartenenza
  /chats/{chat_id}/participants/me:
    parameters:
        - name: chat_id
          in: path
          description: ID univoco della chat.
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
            pattern: '^[a-zA-Z0-9_-]{1,64}$'
    delete:
      tags:
        - gruppo
      summary: Lasciare il gruppo di appartenenza
      description: |- 
        Permette all'utente autenticato di lasciare il gruppo di appartenenza.
      operationId: leaveGroup 
      security: 
        - bearerAuth: []
      responses:
        '200':
          description: Uscita dal gruppo avvenuta con successo
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/Chat'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # SEZIONE 3 (All'interno del path della conversazione) - I Messaggi, essi per costruzione, vivono all'interno di una Chat
...
components:
...
  schemas:
...
    IndividualChat:
      allOf:
        - $ref: '#/components/schemas/ChatBase'
        - type: object
          description: Schema per le chat individuali
          properties:
            type:
              type: string
              enum:
                - individual
              description: Tipo di chat individuale
    
    # Seconda sottoclasse: Chat di gruppo
    GroupChat:
      allOf:
        - $ref: '#/components/schemas/ChatBase'
        - type: object
          description: Schema per le chat di gruppo
          required: 
            - type
            - admins
            - participants
            - created_at
            - updated_at
          properties:
            type:
              type: string
              enum:
                - group
              description: Tipo di chat di gruppo
            group_name:
              type: string
              minLength: 1
              maxLength: 50
              description: Nome della chat di gruppo
              pattern: '^[\w\s]{1,50}$'
              example: "Amici di Universita"
            description:
              type: string
              minLength: 1
              maxLength: 255
              description: Descrizione della chat di gruppo
              pattern: '^[\w\s.,!?-]{1,255}$'
              example: "Chat per organizzare gli eventi universitari"
            admins:
              type: array
              description: Elenco degli amministratori della chat di gruppo
              items:
                type: string
                pattern: '^[a-zA-Z0-9_-]{6,64}$'
                minLength: 1
                maxLength: 64
                description: Elenco degli ID degli utenti amministratori della chat di gruppo
              minItems: 1
              maxItems: 10
            group_icon:
              type: string
              format: uri
              description: URL dell'icona della chat di gruppo
              example: "http://example.com/icons/group123.png"
              minLength: 1
              maxLength: 2048
              pattern: '^https?://.+$'

    # lo schema generico: una chat può essere individuale o di gruppo
    Chat:
      description: Rappresenta una chat (individuale o di gruppo)
      oneOf:
        - $ref: '#/components/schemas/IndividualChat'
        - $ref: '#/components/schemas/GroupChat'
      discriminator:
        propertyName: type
        mapping:
          individual: '#/components/schemas/IndividualChat'
          group: '#/components/schemas/GroupChat'


    # La parte di Input per la creazione di una chat
    CreateIndividualChatRequest:
      type: object
      description: Rappresenta la richiesta di creazione di una chat individuale
      required:
        - type
        - participants
      properties:
        type:
          type: string
          enum:
            - individual
          description: Tipo di chat individuale
        participants:
          type: array
          items:
            type: string
            pattern: '^[a-zA-Z0-9_-]{1,64}$'
            minLength: 1
            maxLength: 64
          minItems: 2
          maxItems: 2 
          description: Elenco degli ID degli utenti partecipanti alla chat (deve contenere esattamente 2 utenti)
    CreateGroupChatRequest:
      type: object
      description: Rappresenta la richiesta di creazione di una chat di gruppo
      required:
        - type
        - participants
        - group_name
      properties:
        type:
          type: string
          enum:
            - group
          description: Tipo di chat di gruppo
        participants:
          type: array
          items:
            type: string
            pattern: '^[a-zA-Z0-9_-]{1,64}$'
            minLength: 1
            maxLength: 64
          minItems: 3
          maxItems: 256
          description: Elenco degli ID degli utenti partecipanti alla chat (minimo 3)
        group_name:
          type: string
          minLength: 1
          maxLength: 50
          description: Nome della chat di gruppo
          pattern: '^[\w\s]{1,50}$'
          example: "Amici di Universita"
        description:
          type: string
          minLength: 1
          maxLength: 255
          description: Descrizione della chat di gruppo
          pattern: '^[\w\s.,!?-]{1,255}$'
          example: "Chat per organizzare gli eventi universitari"
        group_icon:
          type: string
          format: uri
          description: URL dell'icona della chat di gruppo
          minLength: 1
          maxLength: 2048
          pattern: '^https?://.+$'
    CreateChatRequest:
      description: Rappresenta la richiesta di creazione di una nuova chat (individuale o di gruppo)
      oneOf:
        - $ref: '#/components/schemas/CreateIndividualChatRequest'
        - $ref: '#/components/schemas/CreateGroupChatRequest'
      discriminator:
        propertyName: type
        mapping:
          individual: '#/components/schemas/CreateIndividualChatRequest'
          group: '#/components/schemas/CreateGroupChatRequest'

    # Creo gli schemi per i messaggi
...
    SetGroupNameRequest:
      type: object
      required:
        - group_name
      description: Richiesta per impostare o aggiornare il nome della chat di gruppo
      properties:
        group_name:
          type: string
          minLength: 1
          maxLength: 50
          description: Nuovo nome della chat di gruppo
          pattern: '^[\w\s]{1,50}$'
          example: "Nuovo Nome Gruppo"
    SetGroupPhotoRequest:
      type: object
      required:
        - group_icon
      description: Richiesta per impostare o aggiornare la foto della chat di gruppo
      properties:
        group_icon:
          type: string
          format: uri
          description: URL della nuova foto della chat di gruppo
          example: "http://example.com/icons/newgroupicon.png"
          minLength: 1
          maxLength: 500
          pattern: '^https?://.+$'
    SetGroupDescriptionRequest:
      type: object
      required:
        - description
      description: Richiesta per impostare o aggiornare la descrizione della chat di gruppo
      properties:
        description:
          type: string
          minLength: 0
          maxLength: 255
          description: Nuova descrizione della chat di gruppo
          pattern: '^[\w\s.,!?-]{0,255}$'
          example: "Gruppo ufficiale del calcetto del giovedi"
    AddParticipantsRequest:
      type: object
      description: Richiesta per aggiungere membri a una chat di gruppo
      properties:
        user_ids:
          type: array
          items:
            type: string
            pattern: '^[a-zA-Z0-9_-]{6,64}$'
            minLength: 1
            maxLength: 64
          example: ["user123", "user456"]
          minItems: 1
          maxItems: 50
          description: Elenco degli ID degli utenti da aggiungere al gruppo
...
    Error:
      type: object
      description: Rappresenta un errore generico restituito dalle API
      properties:
        code:
          type: string
          minLength: 3
          maxLength: 64
          pattern: '^[A-Z_]{3,64}$'
          example: "NOT_FOUND"
          description: Codice dell'errore
        message:
          type: string
          minLength: 3
          maxLength: 64
          pattern: '^[\s\S]{3,64}$'
          example: "La risorsa richiesta non è stata trovata."
          description: Descrizione dettagliata dell'errore

  # RESPONSES

  # Risposte per gli errori
  responses:
    BadRequest:
      description: Richiesta non valida (400 Bad Request)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Accesso non autorizzato (403 Forbidden)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Risorsa non trovata (404 Not Found)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Errore interno del server (500 Internal Server Error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # PARAMETERS
  parameters:
...
    Ordinamento:
      name: order
      in: query
      description: Criterio di ordinamento (ascendente o discendente).
      required: false
      schema:
        type: string
        enum:
          - asc
          - desc
        default: desc

  # SECURITY SCHEMES
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT